+++ config/nvim/after/ftplugin/javascript.lua
