+++ config/nvim/init.lua
@@ -1,3 +1,4 @@
 require("config.base")
 require("config.lazy")
 require("config.keymap")
+require("config.format")
