+++ config/nvim/lua/plugins/util.lua
@@ -4,13 +4,14 @@ return {
 		event = "VeryLazy",
 		cmd = "Telescope",
 		dependencies = {
-			"nvim-lua/plenary.nvim"
+			"nvim-lua/plenary.nvim",
+			"nvim-telescope/telescope-ui-select.nvim",
 		},
 		config = function()
 			local telescope = require("telescope")
 			telescope.setup({
 				defaults = {
-					layoout_config = {
+					layout_config = {
 						vertical = { width = 0.8 },
 					},
 					mappings = {
@@ -23,8 +24,13 @@ return {
 							-- your custom insert mode mappings
 						},
 					},
+					["ui-select"] = {
+						require("telescope.themes").get_dropdown({})
+						-- even more opts
+					}
 				},
 			})
+			telescope.load_extension("ui-select")
 		end
 	}
 }
