--- config/tmp/nvim/nvim/init.lua
@@ -1,4 +1,0 @@
-require("config.base")
-require("config.lazy")
-require("config.keymap")
-require("config.format")
