--- config/tmp/nvim/nvim/lua/plugins/git.lua
@@ -1,14 +1,0 @@
-return {
-  {
-    "lewis6991/gitsigns.nvim",
-    opts = {
-      signs = {
-        delete = { text = "_" },
-      },
-      signs_staged = {
-        delete = { text = "_" },
-      },
-      numhl = true,
-    },
-  },
-}
