--- config/tmp/nvim/nvim/lua/plugins/ui.lua
@@ -1,195 +1,0 @@
-return {
-  -- the colorscheme should be available when starting Neovim
-  {
-    "folke/tokyonight.nvim",
-    -- lazy = false, -- make sure we load this during startup if it is your main colorscheme
-    priority = 999,   -- make sure to load this before all the other start plugins
-    config = function()
-      -- load the colorscheme here
-      vim.cmd([[colorscheme tokyonight]])
-    end,
-  },
-  --- bufffer line
-  {
-    "akinsho/bufferline.nvim",
-    event = "VeryLazy",
-  },
-  --- status line
-  {
-    "nvim-lualine/lualine.nvim",
-    dependencies = {
-      'nvim-tree/nvim-web-devicons'
-    },
-    event = "VeryLazy",
-    opts = function()
-      local opts = {
-        theme = "auto",
-      }
-      return opts
-    end,
-  },
-  -- tree
-  {
-    "nvim-neo-tree/neo-tree.nvim",
-    branch = "v3.x",
-    dependencies = {
-      "nvim-lua/plenary.nvim",
-      "nvim-tree/nvim-web-devicons",     -- not strictly required, but recommended
-      "MunifTanjim/nui.nvim",
-      -- "3rd/image.nvim", -- Optional image support in preview window: See `# Preview Mode` for more information
-    },
-    keys = {
-      {
-        "<C-e>",
-        function()
-          require("neo-tree.command").execute({ toggle = true })
-        end,
-        desc = "Explorer NeoTree (Root Dir)",
-      },
-    },
-    opts = {
-      window = {
-        width = 25,
-        mappings = {
-          ["<space>"] = "",
-          ["h"] = {
-            "toggle_node",
-            nowait = true,
-          },
-          ["l"] = "open",
-          ["<cr>"] = "",
-          ["hs"] = "open_split",
-          ["vs"] = "open_vsplit",
-          ["s"] = "focus_preview",
-          ["S"] = "",
-        },
-      },
-      filesystem = {
-        window = {
-          fuzzy_finder_mappings = {
-            ["j"] = "move_cursor_down",
-            ["k"] = "move_cursor_up",
-          },
-        },
-      },
-    },
-  },
-  -- which key
-  {
-    "folke/which-key.nvim",
-    event = "VeryLazy",
-  },
-  -- noice
-  {
-    "folke/noice.nvim",
-    event = "VeryLazy",
-    opts = {
-      -- add any options here
-      lsp = {
-        -- override markdown rendering so that **cmp** and other plugins use **Treesitter**
-        override = {
-          ["vim.lsp.util.convert_input_to_markdown_lines"] = true,
-          ["vim.lsp.util.stylize_markdown"] = true,
-          ["cmp.entry.get_documentation"] = true, -- requires hrsh7th/nvim-cmp
-        },
-      },
-      -- you can enable a preset for easier configuration
-      presets = {
-        bottom_search = true,      -- use a classic bottom cmdline for search
-        command_palette = true,    -- position the cmdline and popupmenu together
-        long_message_to_split = true, -- long messages will be sent to a split
-        inc_rename = false,        -- enables an input dialog for inc-rename.nvim
-        lsp_doc_border = false,    -- add a border to hover docs and signature help
-      },
-    },
-    dependencies = {
-      -- if you lazy-load any plugin below, make sure to add proper `module="..."` entries
-      "MunifTanjim/nui.nvim",
-      -- OPTIONAL:
-      --   `nvim-notify` is only needed, if you want to use the notification view.
-      --   If not available, we use `mini` as the fallback
-      "rcarriga/nvim-notify",
-    },
-    init = function()
-      --
-      vim.opt.lazyredraw = false
-    end
-  },
-  -- icons
-  {
-    "echasnovski/mini.icons",
-    lazy = true,
-    opts = {
-      file = {
-        [".keep"] = { glyph = "󰊢", hl = "MiniIconsGrey" },
-        ["devcontainer.json"] = { glyph = "", hl = "MiniIconsAzure" },
-      },
-      filetype = {
-        dotenv = { glyph = "", hl = "MiniIconsYellow" },
-      },
-    },
-    init = function()
-      package.preload["nvim-web-devicons"] = function()
-        require("mini.icons").mock_nvim_web_devicons()
-        return package.loaded["nvim-web-devicons"]
-      end
-    end,
-  },
-  -- TODO:
-  {
-    'b0o/incline.nvim',
-    config = function()
-      local helpers = require("incline.helpers")
-      local devicons = require("nvim-web-devicons")
-      require('incline').setup {
-        render = function(props)
-          local filename = vim.fn.fnamemodify(vim.api.nvim_buf_get_name(props.buf), ':t')
-          if filename == '' then
-            filename = '[No Name]'
-          end
-          local ft_icon, ft_color = devicons.get_icon_color(filename)
-          local function get_git_diff()
-            local icons = { removed = '', changed = '', added = '' }
-            local signs = vim.b[props.buf].gitsigns_status_dict
-            local labels = {}
-            if signs == nil then
-              return labels
-            end
-            for name, icon in pairs(icons) do
-              if tonumber(signs[name]) and signs[name] > 0 then
-                table.insert(labels, { icon .. signs[name] .. ' ', group = 'Diff' .. name })
-              end
-            end
-            if #labels > 0 then
-              table.insert(labels, { '┊ ' })
-            end
-            return labels
-          end
-          local function get_diagnostic_label()
-            local icons = { error = '', warn = '', info = '', hint = '' }
-            local label = {}
-            for severity, icon in pairs(icons) do
-              local n = #vim.diagnostic.get(props.buf, { severity = vim.diagnostic.severity[string.upper(severity)] })
-              if n > 0 then
-                table.insert(label, { icon .. n .. ' ', group = 'DiagnosticSign' .. severity })
-              end
-            end
-            if #label > 0 then
-              table.insert(label, { '┊ ' })
-            end
-            return label
-          end
-          return {
-            { get_diagnostic_label() },
-            { get_git_diff() },
-            { ft_icon and { ' ', ft_icon, ' ', guibg = ft_color, guifg = helpers.contrast_color(ft_color) } or '' },
-            ' ',
-            { filename .. ' ', gui = vim.bo[props.buf].modified and 'bold,italic' or 'bold' },
-          }
-        end
-      }
-    end,
-    -- Optional: Lazy load Incline
-    event = 'VeryLazy',
-  },
-}
